TEST PROJECT
This test project helps to determine the line of json code which caused
problem with a mongoimport command. Actually the mongoimport tool only
gives us the object number in error not the line where it started.
This program should be used only when the file has been produced in
pretty format, otherwise is not nececessary because the number object is
the current line in error